+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 0.6.20
 ******
 * Cleaned up how PES descriptors are generated & stored in the ``ARMC()`` class.
+* Atom names specified in .PSF files are now accessible by ``MultiMolecule()`` instances
+  during the ARMC procedure.
 
 
 0.6.19