+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 * Cleaned up how PES descriptors are generated & stored in the ``ARMC()`` class.
 * Atom names specified in .PSF files are now accessible by ``MultiMolecule()`` instances
   during the ARMC procedure.
+* Generalized ``dekekulize()`` to work for all non-integer bond orders; not just ``1.5``.
 
 
 0.6.19